+++ Documents/ReleaseNotes.md
@@ -4,6 +4,7 @@
 - `11.0.0-Preview-1.6`: 
   - `Data-Driven Styles`.
 	- `MatchStyleBuilder`. A new builder for creating match-style expressions.
+	- `LinearInterpolateStyleBuilder`. A new builder for creating linear interpolate-style expressions.
 - `11.0.0-Preview-1.5`: 
   - `Layer Options`. Any remaining layer option properties that support data-driven styling have been updated to support both direct values and expressions.
 - `11.0.0-Preview-1.4`: